+++ src/main/java/com/example/Balenz/article/dto/SimpleArticleDto.java
@@ -16,4 +16,6 @@ public class SimpleArticleDto {
 
     private FrameType frameType;
 
+    private String imageUrl;
+
 }
